More accurately rename mirage scenarios -> fixtures
