+++ test/integration/mirage-context-resolvers.test.ts
@@ -2,7 +2,7 @@ import { expect } from 'chai';
 import { graphqlSchema } from './test-helpers/test-schema';
 import defaultResolvers from './test-helpers/mirage-static-resolvers';
 import { server as mirageServer } from './test-helpers/mirage-sample';
-import defaultScenario from './test-helpers/mirage-sample/scenarios/default';
+import defaultScenario from './test-helpers/mirage-sample/fixtures';
 import { createQueryHandler } from '../../src/graphql';
 
 const { query: graphQLHandler } = createQueryHandler(defaultResolvers, {
